- Added some comments in the code.
- It changed the name of pizza-detalis to pizza-details component.
- Added the topping selection functionality for add on the pizza.

+++ Frontend/src/app/pages/topping/list-toppings.component.ts
@@ -4,6 +4,8 @@ import { Topping } from 'src/app/models/topping.model';
 import { ActivatedRoute, Router } from '@angular/router';
 import { PizzaService } from 'src/app/services/pizza.service';
 
+import Swal from 'sweetalert2';
+
 @Component({
   selector: 'app-list-toppings',
   templateUrl: './list-toppings.component.html',
@@ -14,46 +16,74 @@ export class ListToppingsComponent implements OnInit {
   toppingsList: Topping[] = [];
   id: string = '';
 
+  // Declaration of messages
+  Toast = Swal.mixin({
+    toast: true,
+    position: 'top',
+    showConfirmButton: false,
+    timer: 3000
+  });
+
   constructor(private toppingService: ToppingService,
     private pizzaService: PizzaService,
     private activatedRoute: ActivatedRoute,
     private router: Router) {
-    activatedRoute.params.subscribe(params => {
-      console.log(params.id);
-      this.id = params.id
-    });
+
+    // Getting the topping id from router
+    activatedRoute.params.subscribe(params => this.id = params.id );
   }
 
   ngOnInit() {
+    // This method verify if the id exists for load the list or available 
+    // list of toppings for the pizza id
     if (this.id)
       this.loadAvailableList();
     else
       this.loadList();
   }
 
+  // Load the list of toppings
   loadList() {
     this.toppingService.getToppings()
       .subscribe(toppingsList => {
         this.toppingsList = toppingsList;
-        console.log(this.toppingsList);
       });
   }
 
+  // Load the available list of toppings for the pizza id
   loadAvailableList() {
     this.pizzaService.getAvailableToppingsForPizza(this.id)
       .subscribe(toppingsList => {
         this.toppingsList = toppingsList;
-        console.log(this.toppingsList);
       });
   }
 
   deleteTopping(topping: Topping) {
-    this.toppingService.deleteTopping(topping).subscribe(resp => {
-      console.log(resp)
-      this.loadList();
+
+    Swal.fire({
+      title: 'Do you want to delete this topping?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+      if (result.value) {
+        // Deleting topping
+        this.toppingService.deleteTopping(topping).subscribe(resp => {
+          this.loadList();
+        });
+        //Message
+        this.Toast.fire({
+          icon: 'info',
+          title: 'Deleted successfully'
+        })
+      }
     });
   }
 
+  // It select the topping and is add to the pizza
   selectTopping(topping: Topping) {
     this.pizzaService.addToppingToPizza(topping, this.id).subscribe(resp => {
       this.router.navigate(['/pizza-details', this.id]);
